Add topics; make resources and topics translatable
